*** empty log message ***

+++ Berlin/include/Berlin/Math.hh
@@ -37,8 +37,8 @@ public:
 
   static float abs(float a) { return fabs(a);}
   static double abs(double a) { return fabs(a);}
-  static float abs(int a) { return abs(a);}
-  static double abs(long a) { return abs(a);}
+  static int abs(int a) { return abs(a);}
+  static long abs(long a) { return abs(a);}
 
   template <class T> static int round(T a) { return a > 0 ? static_cast<int>(a + 0.5) : - static_cast<int>(-a + 0.5);}
